% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_repo.R
\name{adaptRpkgs}
\alias{adaptRpkgs}
\title{adapts packages in Rlibs according to todo dataframe
  1) removes package versions in Rlibs that are not in todo
  2) installs all missing package versions in todo}
\usage{
adaptRpkgs(todo, Rlibs = NULL, installlib = Rlibs)
}
\arguments{
\item{todo}{input dataframe of packages to change}

\item{Rlibs}{character vector of R libraries to modify packages.
Default NULL means .libPaths() is read.}

\item{installlib}{library to install packages to; defaults to first element of Rlibs.}
}
\value{
rc named vector of installation return codes.
}
\description{
adapts packages in Rlibs according to todo dataframe
  1) removes package versions in Rlibs that are not in todo
  2) installs all missing package versions in todo
}
