% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepkgs.R
\docType{data}
\name{basepkgs}
\alias{basepkgs}
\title{List of base R packages.}
\format{A data frame with 14 rows and 3 variables:
\describe{
  \item{Package}{Package name}
  \item{Version}{Package version, effectively the R version}
  \item{Priority}{Package priority, here always 'base'}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
basepkgs
}
\description{
A dataset containing the package names, version, and priority.
}
\keyword{datasets}
