% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baserecpkgs.R
\docType{data}
\name{baserecpkgs}
\alias{baserecpkgs}
\title{List of base and recommended R packages.}
\format{A data frame with 14 rows and 3 variables:
\describe{
  \item{Package}{Package name}
  \item{Priority}{Package priority, base or recommended}
}}
\usage{
baserecpkgs
}
\description{
A dataset containing the package names, version, and priority.
}
\keyword{datasets}
