% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_repo.R
\name{compareRmodel}
\alias{compareRmodel}
\title{compares current R installation in Rlibs with Rmodel}
\usage{
compareRmodel(Rmodel, Rlibs)
}
\arguments{
\item{Rmodel}{location of the model R repo}

\item{Rlibs}{character vector of R libraries to check}
}
\value{
dataframe with packages to change
}
\description{
compares current R installation in Rlibs with Rmodel
}
