% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_repo.R
\name{getRpkgs}
\alias{getRpkgs}
\title{retrieve current R package setup}
\usage{
getRpkgs(Rlibs = NULL)
}
\arguments{
\item{Rlibs}{character vector of R libraries to check}
}
\value{
dataframe with 3 variables: Package, Version, and Priority
}
\description{
retrieve current R package setup
}
