% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_repo.R
\name{list_src_contrib}
\alias{list_src_contrib}
\title{list packages provided in srccontrib directory}
\usage{
list_src_contrib(srccontrib = "src/contrib")
}
\arguments{
\item{srccontrib}{directory of source R package bundles}
}
\value{
dataframe sorted by package name
}
\description{
list packages provided in srccontrib directory
}
